+++ lib/nanoc-conref-fs.rb
@@ -1,4 +1,5 @@
 require 'nanoc-conref-fs/ancestry'
-require 'nanoc-conref-fs/conref-fs'
 require 'nanoc-conref-fs/conrefifier'
 require 'nanoc-conref-fs/datafiles'
+require 'nanoc-conref-fs/variables'
+require 'nanoc-conref-fs/conref-fs'
