+++ lib/nanoc-conref-fs/variables.rb
@@ -1,0 +1,29 @@
+module NanocConrefFS
+  # Unsure why attr_accessor does not work here
+  module Variables
+    def self.variables
+      @variables
+    end
+
+    def self.variables=(variables)
+      @variables = variables
+    end
+
+    def self.data_files
+      @data_files
+    end
+
+    def self.data_files=(data_files)
+      @data_files = data_files
+    end
+
+    def self.fetch_data_file(association)
+      reference = association.split('.')
+      data = variables['site']['data']
+      while key = reference.shift
+        data = data[key]
+      end
+      data
+    end
+  end
+end
