+++ test/fixtures/layouts/retrieve.html
@@ -8,7 +8,7 @@
 
 <body>
 
-  <%= VariableMixin.variables['site']['data']['variables']['product']['product_name']['2.0'] %>
+  <%= NanocConrefFS::Variables.variables['site']['data']['variables']['product']['product_name']['2.0'] %>
   <%= yield %>
 
 </body>
